requirements.txt
streamlit
rank-bm25 
sentence-transformers 
numpy 
scipy 
openai 
pdfplumber 
python-docx 
beautifulsoup4 
requests
